Rename dragActivities folder to games
