+++ src/BloomBrowserUI/bookEdit/toolbox/games/GamePromptDialog.tsx
@@ -21,7 +21,7 @@ import {
     adjustTarget,
     DragActivityTool,
     makeTargetForDraggable
-} from "./GameTool";
+} from "../games/GameTool";
 import * as ReactDOM from "react-dom";
 import BloomSourceBubbles from "../../sourceBubbles/BloomSourceBubbles";
 import {
